feat: 기업 상세조회 common으로 이동
#61
