+++ Application/Reward/src/main/java/com/example/Reward/Common/Dto/DetailEnterPriseResponseDto.java
@@ -1,4 +1,4 @@
-package com.example.Reward.Advertisement.Dto.out;
+package com.example.Reward.Common.Dto;
 
 import com.example.Reward.Common.Entity.Event;
 import lombok.Builder;
@@ -9,13 +9,9 @@ import lombok.Getter;
 public class DetailEnterPriseResponseDto {
 
     private String name;
-
     private String comment;
 
-
-
     public static DetailEnterPriseResponseDto of(Event event) {
-
         return DetailEnterPriseResponseDto.builder().name(event.getEnterpriseName()).comment(event.getComment()).build();
     }
 }
