+++ Application/Reward/src/main/java/com/example/Reward/Common/Repository/EventRepository.java
@@ -1,6 +1,7 @@
 package com.example.Reward.Common.Repository;
 
 import com.example.Reward.Common.Entity.Event;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
@@ -16,4 +17,6 @@ public interface EventRepository extends JpaRepository<Event,Long> {
     List<Event> findByRewardAmountLessThan(Long rewardAmount);
 
     List<Event> findByRewardAmountGreaterThanEqual(Long rewardAmount);
+
+    List<Event> findByEnterpriseName(String enterpriseName, Pageable pageable);
 }
